Add xp-gateway-bitcoin
