+++ primitives/gateway/bitcoin/Cargo.toml
@@ -1,0 +1,44 @@
+[package]
+name = "xp-gateway-bitcoin"
+version = "2.0.0-alpha.4"
+authors = ["The ChainX Authors"]
+edition = "2018"
+
+[dependencies]
+codec = { package = "parity-scale-codec", version = "1.3.4", default-features = false, features = ["derive"] }
+hex = { version = "0.4", default-features = false }
+serde = { version = "1.0", features = ["derive"], optional = true }
+
+# Substrate primitives
+sp-core = { version = "2.0.0", default-features = false }
+sp-runtime = { version = "2.0.0", default-features = false }
+sp-std = { version = "2.0.0", default-features = false }
+frame-support = { version = "2.0.0", default-features = false }
+
+# ChainX primitives
+chainx-primitives = { path = "../../../primitives", default-features = false }
+xp-gateway-common = { path = "../../gateway/common", default-features = false }
+xp-logging = { path = "../../logging", default-features = false }
+
+# light-bitcoin
+light-bitcoin = { git = "https://github.com/chainx-org/light-bitcoin", branch = "master", default-features = false }
+
+[features]
+default = ["std"]
+std = [
+    "codec/std",
+    "hex/std",
+    "serde",
+    # Substrate primitives
+    "sp-core/std",
+    "sp-runtime/std",
+    "sp-std/std",
+    "frame-support/std",
+    # ChainX primitives
+    "chainx-primitives/std",
+    "xp-gateway-common/std",
+    "xp-logging/std",
+    # light-bitcoin
+    "light-bitcoin/std",
+]
+ss58check = ["xp-gateway-common/ss58check"]
