+++ primitives/gateway/bitcoin/src/extractor.rs
