+++ runtime/Cargo.toml
@@ -67,6 +67,7 @@ orml-currencies = { version = "0.3", default-features = false }
 
 # ChainX primitives
 chainx-primitives = { path = "../primitives", default-features = false }
+xp-gateway-bitcoin = { path = "../primitives/gateway/bitcoin", default-features = false }
 xp-logging = { path = "../primitives/logging", default-features = false }
 xp-mining-staking = { path = "../primitives/mining/staking", default-features = false }
 xp-protocol = { path = "../primitives/protocol", default-features = false }
@@ -150,6 +151,7 @@ std = [
     "orml-currencies/std",
     # ChainX primitives
     "chainx-primitives/std",
+    "xp-gateway-bitcoin/std",
     "xp-logging/std",
     "xp-mining-staking/std",
     "xp-protocol/std",
