+++ xpallets/gateway/bitcoin/Cargo.toml
@@ -28,6 +28,7 @@ orml-utilities = { version = "0.3", default-features = false }
 # ChainX primitives
 chainx-primitives = { path = "../../../primitives", default-features = false }
 xp-gateway-common = { path = "../../../primitives/gateway/common", default-features = false }
+xp-gateway-bitcoin = { path = "../../../primitives/gateway/bitcoin", default-features = false }
 xp-logging = { path = "../../../primitives/logging", default-features = false }
 xp-protocol = { path = "../../../primitives/protocol", default-features = false }
 
@@ -70,6 +71,7 @@ std = [
     # ChainX primitives
     "chainx-primitives/std",
     "xp-gateway-common/std",
+    "xp-gateway-bitcoin/std",
     "xp-logging/std",
     "xp-protocol/std",
     # ChainX pallets
@@ -85,4 +87,5 @@ runtime-benchmarks = [
 ]
 ss58check = [
     "xp-gateway-common/ss58check",
+    "xp-gateway-bitcoin/ss58check",
 ]
